Separate leo and tpi executables (#662)
* Separate leo and tpi.

* Update goreleaser config to build leo and tpi separately.

* Add dist to .gitignore.

* Use leo name.

* Minor Makefile update.

* Update machine script.

* Update golden files.

* Keep terraform-provider-iterative in package root.

* Keep using existing tpi release in machine script.

* Update golden files.

* Address review comments.
