+++ .gitignore
@@ -7,6 +7,9 @@ bin
 /*.tfstate*
 /crash.log
 
+# Build artefacts
+/dist
+
 # Ignore any .tfvars files that are generated automatically for each Terraform run. Most
 # .tfvars files are managed as part of configuration and so should be included in
 # version control.
