+++ cmd/leo/main.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"os"
+)
+
+func main() {
+	cmd := NewCmd()
+	err := cmd.Execute()
+	if err != nil {
+		os.Exit(1)
+	}
+}
