+++ main.go
@@ -1,24 +1,15 @@
 package main
 
 import (
-	"os"
-
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 
-	"terraform-provider-iterative/iterative/utils"
 	"terraform-provider-iterative/iterative"
-	"terraform-provider-iterative/cmd"
+	"terraform-provider-iterative/iterative/utils"
 )
 
 func main() {
-  defer utils.WaitForAnalyticsAndHandlePanics()
-  
-	if os.Getenv(plugin.Handshake.MagicCookieKey) != plugin.Handshake.MagicCookieValue {
-		cmd.Execute()
-		return
-	}  
-
+	defer utils.WaitForAnalyticsAndHandlePanics()
 	plugin.Serve(&plugin.ServeOpts{
 		ProviderFunc: func() *schema.Provider {
 			return iterative.Provider()
